projeto6/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asyncio==4.0.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
dataclasses-json==0.6.7
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-ollama==1.0.1
langchain-text-splitters==1.1.0
langsmith==0.6.4
markdown-it-py==4.0.0
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.4.1
ollama==0.6.1
orjson==3.11.5
packaging==25.0
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
setuptools==80.9.0
SQLAlchemy==2.0.46
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
wheel==0.45.1
yarl==1.22.0
zstandard==0.25.0
